Add CI for CTest and clang-tidy (#11)
* Add CI runs for Clang Tidy

* Use ARM for GitHub runners

* Disable building of sorting homework due to assembly architecture issues

* Run tests using CTest

* Add .clang-tidy

* Update clang-tidy

* Fix basic clang-tidy warnings

* Attempt to fix uninitialized value warning

* Include assert

* Use calloc for allocation

* Use operators instead of operator

* Fix clang-tidy warnings

* Format

* Fix char rounding

* Fix job name

+++ hws/paren/main.c
@@ -9,8 +9,8 @@ int main(void)
     int n = 0;
     bool ok = true;
 
-    int len = strlen(input);
-    for (int i = 0; i < len; i++) {
+    size_t len = strlen(input);
+    for (size_t i = 0; i < len; i++) {
         if (input[i] == '(') {
             n += 1;
         } else if (input[i] == ')') {
